pyproject.toml
[project]
name = "contractor"
version = "0.1.0"
description = "Use llm to generate openapi specification from code"
authors = [
    {name = "grauwolf32"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "google-adk (>=1.21.0,<2.0.0)",
    "litellm (>=1.80.12,<2.0.0)",
    "langfuse (>=3.11.2,<4.0.0)",
    "openinference-instrumentation-google-adk (>=0.1.8,<0.2.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[[tool.poetry.source]]
name = "main-artifactory"
url = "https://artifactory.tcsbank.ru/artifactory/api/pypi/python-all/simple/"
priority = "primary"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
